assistant: Restructure ambient context in preparation for adding more
