+++ crates/assistant/src/ambient_context.rs
@@ -1,0 +1,8 @@
+mod recent_buffers;
+
+pub use recent_buffers::*;
+
+#[derive(Default)]
+pub struct AmbientContext {
+    pub recent_buffers: RecentBuffersContext,
+}
